requirements.txt
aiocontextvars==0.2.2
appdirs==1.4.4
asgiref==3.2.10
attrs==19.3.0
backcall==0.2.0
black==19.10b0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3
colorlog==4.2.1
contextvars==2.4
coreapi==2.3.3
coreschema==0.0.4
decorator==4.4.2
Django==2.2.15
django-cors-headers==3.4.0
django-filter==2.3.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
drf-yasg==1.17.1
filetype==1.0.7
har2case==0.3.1
httprunner==2.3.0
idna==2.10
immutables==0.14
importlib-metadata==1.7.0
inflection==0.5.0
ipython==7.16.1
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.9.5
jsonpath==0.82
loguru==0.4.1
Markdown==3.2.2
MarkupSafe==1.1.1
more-itertools==8.4.0
mysqlclient==2.0.1
packaging==20.4
parso==0.7.1
pathspec==0.8.0
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.5
ptyprocess==0.6.0
py==1.9.0
pydantic==1.6.1
Pygments==2.6.1
PyJWT==1.7.1
pyparsing==2.4.7
pytest==5.4.3
pytest-html==2.1.1
pytest-metadata==1.10.0
pytz==2020.1
PyYAML==5.3.1
regex==2020.7.14
requests==2.24.0
requests-toolbelt==0.9.1
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
sentry-sdk==0.14.4
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
traitlets==4.3.3
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.10
wcwidth==0.2.5
zipp==3.1.0
